# openapi.yml
...
openapi: 3.0.0
info:
  title: Dre Sean Sutton's Smash Ultimate API
  description: >
    This is *the* Smash Ultimate restful API that offers indepth
    frame data information on the cast. In GET requests, use a query parameter to specify
    a single Fighter. Values separated by `/`, `()`, and `[]`
    indicate different hitboxes that are mirrored
    by the `moveType` property. All frame data is copied from the
    fantastic [Ultimate Frame Data](https://ultimateframedata.com)
    website, occasionally modified for clarity, and manually input
    into my database. If you see any mistakes, please let me know on GitHub!
    </br>
    </br>
    80 Fighters, so far, and they are frequently added!
    </br>
    </br>
    [GitHub](https://github.com/DreSean-Sutton/the-ultimate-api)
    [LinkedIn](https://www.linkedin.com/in/dre-sean-sutton/)

  version: v9
paths:
  /api/get/fighters:
    get:
      tags:
        - fighters
      summary: Returns basic information for all fighters
      description: >
       Use `fighter`, `fighterId`, and `rosterId`
       to specify a single fighter
      parameters:
      - in: query
        name: fighter
        schema:
          type: string
        required: false
      - in: query
        name: fighterId
        schema:
          type: integer
        required: false
      - in: query
        name: rosterId
        schema:
          type: integer
        required: false
      - in: query
        name: orderByRosterId
        schema:
          type: boolean
        required: false
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    figherId:
                      type: integer
                    fighter:
                      type: string
                    rosterId:
                      type: integer
                    displayName:
                      type: string
                example:
                -
                  fighterId: 1
                  fighter: banjo
                  rosterId: 73
                  displayName: Banjo & Kazooie
                -
                  fighterId: 2
                  fighter: bayonetta
                  rosterId: 63
                  displayName: Bayonetta
  /api/get/fighters/data:
    get:
      tags:
        - fighters
      summary: Returns all frame data for all Fighters
      description: >
       Use `fighter`, `fighterId`, and `rosterId`
       to specify a single fighter
      parameters:
      - in: query
        name: fighter
        schema:
          type: string
        required: false
      - in: query
        name: fighterId
        schema:
          type: integer
        required: false
      - in: query
        name: rosterId
        schema:
          type: integer
        required: false
      - in: query
        name: orderByRosterId
        schema:
          type: boolean
        required: false
      responses:
        '200':
          description:
            OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/allData'
  /api/get/fighters/data/{type}:
    get:
      tags:
      - fighters
      summary: Returns a specific type of frame data for all Fighters
      description: >
       Use `fighter`, `fighterId`, and `rosterId`
       to specify a single fighter
      parameters:
      - in: path
        name: type
        schema:
          type: string
        required: true
        description: Must be `moves`, `throws`, `movements` or `stats`
      - in: query
        name: fighter
        schema:
          type: string
        required: false
      - in: query
        name: fighterId
        schema:
          type: integer
        required: false
      - in: query
        name: rosterId
        schema:
          type: integer
        required: false
      - in: query
        name: orderByRosterId
        schema:
          type: boolean
        required: false
      responses:
        '200':
          description:
            OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/allData'
...
components:
  schemas:
...
    allData:
      type: array
      items:
        type: object
        properties:
          activeFrames:
            type: string
          category:
            type: string
          damage:
            type: string
          displayName:
            type: string
          fighter:
            type: string
          fighterId:
            type: integer
          moveType:
            type: string
          name:
            type: string
          rosterId:
            type: integer
          statValue:
            type: string
          totalFrames:
            type: string
          type:
            type: string
      example:
      -
        activeFrames: "6/11-25 [rehit: 3]/27-28"
        category: special
        damage: 3.0/0.2/3.0%
        displayName: Bayonetta
        fighter: bayonetta
        fighterId: 2
        moveType: first/multi/final
        name: up b
        rosterId: 69
        totalFrames: 31
        type: move
      -
        activeFrames: 7-8
        damage: null
        displayName: Bayonetta
        fighter: bayonetta
        fighterId: 2
        name: grab
        rosterId: 69
        totalFrames: 37
        type: throw
      -
        activeFrames: 3-29
        displayName: Inkling
        fighter: inkling
        fighterId: 13
        name: neutral air dodge
        rosterId: 70
        totalFrames: 50
        type: movement
      -
        displayName: Inkling
        fighter: inkling
        fightrId: 13
        name: out of shield, neutral air/up smash
        rosterId: 70
        statValue: 9 frames
        type: stat
...